Honesty pass: f64 reference (not long-double), regenerate doc numbers, 3L-exp fix, full overflow series, verify over all scales
- Reference is IEEE double (f64, 53-bit); on Apple Silicon long double == double, so drop 'long-double oracle' wording everywhere (README/claims/PREREG/REVIEW) and state the real 53-vs-24-bit, ~5e8x headroom. Fix REVIEW's false '~64-bit mantissa / 2^-40 / four-billion-times' claim.
- Regenerate README + claims headline numbers to match frontier.md: worst error ratio 1.43 (scale=1 near f32 floor), not fabricated 1.19; regenerate the fabricated scale=1 error column from committed data.
- Latency: reframe flat '~2.8x' as declining ~2.6-2.9x above the timer-resolution floor; short sub-100ns lengths are steady_clock-noisy.
- Correct the exp accounting: online does two exp per element in the reduce pass plus one in normalization (~3L), not ~2L; the 3:1 ratio explains the slowdown (README/claims/analyze.py).
- Overflow: show the FULL non-monotone series (64->12, 128->0, 256->38, 512->25, ...) and say it rises on average, noisily (8 seeds = 12.5% granularity).
- verify.py: compute worst error ratio over ALL scales (reports ~1.43, consistent with analyze.py) and reword the overflow docstring from 'monotone non-decreasing' to reaches 100% by the longest length.

+++ src/bench.cpp
@@ -1,7 +1,7 @@
 // Measure the accuracy and cost of the three softmax variants over a grid of sequence lengths and
 // score magnitudes, with several seeds per cell for confidence intervals. Data is generated in-process
 // with a deterministic PRNG so the benchmark is self-contained and reproducible. For each (length,
-// scale, seed) it emits the max-abs error of each f32 variant against the long-double reference (or
+// scale, seed) it emits the max-abs error of each f32 variant against the double-precision (f64) reference (or
 // "nan" if the variant produced a non-finite result) and the minimum-over-reps latency of the two safe
 // variants.
 //
@@ -76,7 +76,7 @@ int main(int argc, char** argv) {
                 std::vector<float> x(static_cast<size_t>(L));
                 for (float& v : x) v = normal(scale);
 
-                std::vector<long double> ref;
+                std::vector<double> ref;
                 fsm::softmax_reference(x, ref);
                 std::vector<float> tp, on, na;
                 fsm::softmax_twopass(x, tp);
